backend/requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
uuid==1.30
sqlparse==0.4.4
typing-extensions==4.7.1
celery==5.3.1
redis==4.6.0
pytz==2023.3
gunicorn==20.1.0
flower==2.0.0
django-simple-captcha==0.5.18
requests==2.31.0